// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Tue, 27 Jun 2023 16:44:35 GMT
 */

/**
 * @enum
 * @readonly
 * @description TokenWorlds Action Names
 */
export enum TokenWorldsActionName {
    Burn = 'burn',
    Cancel = 'cancel',
    Claimunstkes = 'claimunstkes',
    Close = 'close',
    Create = 'create',
    Issue = 'issue',
    Memberreg = 'memberreg',
    Memberunreg = 'memberunreg',
    Newmemterms = 'newmemterms',
    Stake = 'stake',
    Stakeconfig = 'stakeconfig',
    Staketime = 'staketime',
    Transfer = 'transfer',
    Unlock = 'unlock',
    Unstake = 'unstake',
};